Added GPIO MMIO mapping for BCM2711 mini-UART used by SerialServer.
This wasn't needed before, because the previous mini-UART driver
didn't explicitly configure GPIO UART pins when initializing UART
drivers. Now it does, and BCM2711 GPIO driver was failing because
there was no device untyped for the driver to map to the MMIO registers.

TODO: should this be moved to GPIOMUXServer and have the SerialServer
UART(s) handle GPIO via that?